fix: #WB2-1604, fix feedback PR create a hook blocker + clean code
